Working on showing search results when grabbing from the api.
Potential issues:
Right now, the search function is too simple. When I input a search string, it responds directly to .com/username/repo. This does not allow for a dynamic search experience where you do not necessarily have to input the proper username
in order to find something.

This can easily be fixed by making the search function more advanced. I would need to first search in users, and then grab the best matching username and insert that username into the repos-link.